requirements.txt
torch
torchvision
numpy
pillow
matplotlib
tqdm
pyyaml
mlflow
fastapi
uvicorn
python-multipart
